move the patch to its own library to better clean up linker tricks
