+++ boards/t_watch_s3/CMakeLists.txt
@@ -4,10 +4,3 @@ zephyr_library_sources(
     t_watch_s3_boot.c
 )
 
-# for access to the bma4xx sensor data format
-# to properly hot-patch the sensor decode api
-zephyr_library_include_directories(
-    ${ZEPHYR_BASE}/drivers/sensor/bosch/bma4xx
-)
-
-
