+++ tests/prj.conf
@@ -2,3 +2,5 @@ CONFIG_ZTEST=y
 CONFIG_BRINGUP_LOG_LEVEL_DBG=y
 CONFIG_ZTEST_SHELL=y
 CONFIG_LORAMAC_REGION_US915=y
+CONFIG_BMA4XX_BOOT_RESET_HACK=y
+CONFIG_BMA4XX_DECODER_PATCH_HACK=y
